# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Office Management API
  version: 1.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /admin/departments/getAll:
    get:
      summary: Get all departments with manager and staff count
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of departments
        "500":
          description: Server error
  /admin/departments/get/{id}:
    get:
      summary: Get a single department with staff list
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department details with staff
        "404":
          description: Department not found
        "500":
          description: Server error
  /admin/departments/add:
    post:
      summary: Create a new department
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                manager_id:
                  type: integer
      responses:
        "210":
          description: Department created successfully
        "400":
          description: Department name already exists
        "500":
          description: Server error
  /admin/departments/update/{id}:
    put:
      summary: Update an existing department
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                manager_id:
                  type: integer
      responses:
        "200":
          description: Department updated successfully
        "400":
          description: Department name already taken
        "404":
          description: Department not found
        "500":
          description: Server error
  /admin/departments/delete/{id}:
    delete:
      summary: Delete a department
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department deleted successfully
        "404":
          description: Department not found
        "500":
          description: Server error
  /admin/departments/getPotentialManagers:
    get:
      summary: Get list of potential managers (admins and managers)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of potential managers
        "500":
          description: Server error
...
  /superadmin/departments/getAll:
    get:
      summary: Get all departments with manager and staff count (Super Admin)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of departments
        "500":
          description: Server error
  /superadmin/departments/get/{id}:
    get:
      summary: Get a single department with staff list (Super Admin)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department details with staff
        "404":
          description: Department not found
        "500":
          description: Server error
  /superadmin/departments/add:
    post:
      summary: Create a new department (Super Admin)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                manager_id:
                  type: integer
      responses:
        "210":
          description: Department created successfully
        "400":
          description: Department name already exists
        "500":
          description: Server error
  /superadmin/departments/update/{id}:
    put:
      summary: Update an existing department (Super Admin)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                manager_id:
                  type: integer
      responses:
        "200":
          description: Department updated successfully
        "400":
          description: Department name already taken
        "404":
          description: Department not found
        "500":
          description: Server error
  /superadmin/departments/delete/{id}:
    delete:
      summary: Delete a department (Super Admin)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department deleted successfully
        "404":
          description: Department not found
        "500":
          description: Server error
  /superadmin/departments/getPotentialManagers:
    get:
      summary: Get list of potential managers (Super Admin)
      tags:
        - Admin-Departments
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of potential managers
        "500":
          description: Server error
...